Use relative link
Whilst this doesn't work when you're viewing it from the turtlecoin-wiki repo, once it's copied across to the turtlecoin/wiki section, it will be a relative link to the correct guide - keeping you on the wiki page rather than into a less user friendly standard repo. I suppose we could use absolute links to the turtlecoin/wiki links so it works here as well.